requirements.txt
pyodbc
pycryptodomex